Fixed bug in how time was calculated.  Seems to have cleared up many other issues.

+++ core/src/main/java/com/dumptruckman/darkascendance/shared/messages/Message.java
@@ -58,7 +58,7 @@ public class Message extends MessageBase implements Cloneable {
         return forAllButOneConnections;
     }
 
-    public long getTime() {
+    public float getTime() {
         return messageTime;
     }
